zeroMem after each chrom
closes #12

this bug would only manifest when later chroms were longer than
earlier chroms--this usually does not happen, but can, even
in human for e.g. chrX.

thanks @raul-w for reporting and providing a test-case.

this also changes duphold so that the median chromosomal depth
is calcualted for **non-zero** depths only.

the event depths are calculated using zero depths otherwise we
would lose the ability to detect homozygous deletions.